Web Game - Popular Memory game
This is a web Memory game created in HTML, CSS and JS with basic jQuery functions. 
Goal of the game is to unravel all the cards in as few turns as possible. 
Location of  respective cards is randomized at beginning of each game. 
Player has to guess where are the pairs of the same pictures of famous buildings from all over the world. 
Card is shown when is clicked on. If pair is found in one turn this pair disappear.
If cards don't match they are hidden again after two seconds. Clicking on another cards is disabled during this time. 
Unraveling two cards increase the turn counter.
After finding all pairs game ends and player is asked if wants to play again.